Endringer etter kommentarer i PR fra Lars Otto

+++ src/test/java/no/nav/bidrag/beregn/barnebidrag/BarnebidragCoreTest.java
@@ -87,7 +87,7 @@ public class BarnebidragCoreTest {
             .isEqualTo(LocalDate.parse("2018-01-01")),
 
         () -> assertThat(beregnBarnebidragResultatCore.getResultatPeriodeListe().get(0).getResultatGrunnlag().getSjablonListe().get(0)
-            .getSjablonInnholdListe().get(0).getSjablonInnholdVerdi().doubleValue()).isEqualTo(22)
+            .getSjablonInnholdListe().get(0).getSjablonInnholdVerdi()).isEqualTo(BigDecimal.valueOf(22))
 
     );
   }
